* commit 'c5697e92380b0bdc59394973f4d19c28ca40dafc':
  Add explicit `#include <cstdint>`.
  Enable -Wsign-conversion

+++ tests/testaux/cx_string.hpp
@@ -35,7 +35,7 @@ namespace testaux
 	{
 		auto p = s;
 		while (*p) { ++p; }
-		return p - s;
+		return static_cast<std::size_t>(p - s);
 	}
 
 	static_assert(cx_strlen("") == 0, "");
